% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ARD_dRMST.R
\name{calculate.ARD.dRMST}
\alias{calculate.ARD.dRMST}
\title{Calculate ARD and dRMST}
\usage{
calculate.ARD.dRMST(S = NULL, W = NULL, horizon = 0)
}
\arguments{
\item{S}{the outcome, a Surv() object}

\item{W}{treatment indicator, 1 for treatment and 0 for control}

\item{horizon}{the prediction horizon}
}
\value{
The output of the calculate.ARD.dRMST function is a "list" with the following components.

LRT.p

p-value of the log-rank test


ARD

absolute risk difference


ARD_se

standard error of absolute risk difference


dRMST

difference in restricted mean survival time


dRMST_se

standard error of differnece in restricted mean survival time
}
\description{
This function calculates the absolute risk difference (ARD) and difference in restricted mean survival time (dRMST)
}
\examples{
library(PredictionTools)
library(survival)

# load data
data(pbc, package="survival")

# set survival time
S <- survival::Surv(time=pbc$time/365, event=as.numeric(pbc$status!=0))

# treatment indicator
W <- pbc$trt

# calculate ARD and dRMST
calculate.ARD.dRMST(S=S, W=W, horizon=5)
}
