% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_RD_dRMST.R
\name{calculate.RD.dRMST}
\alias{calculate.RD.dRMST}
\title{Calculate RD and dRMST}
\usage{
calculate.RD.dRMST(S = NULL, W = NULL, horizon = 0)
}
\arguments{
\item{S}{the outcome, a Surv() object}

\item{W}{treatment indicator, 1 for treatment and 0 for control}

\item{horizon}{the prediction horizon}
}
\value{
The output of the calculate.RD.dRMST function is a "list" with the following components.

LRT.p

p-value of the log-rank test


RD

absolute risk difference


RD_se

standard error of absolute risk difference


dRMST

difference in restricted mean survival time


dRMST_se

standard error of differnece in restricted mean survival time
}
\description{
This function calculates the absolute risk difference (RD) and difference in restricted mean survival time (dRMST)
}
\examples{
library(PredictionTools)
library(survival)

# load data
data(pbc, package="survival")

# set survival time
S <- survival::Surv(time=pbc$time/365, event=as.numeric(pbc$status!=0))

# treatment indicator
W <- pbc$trt

# calculate RD and dRMST
calculate.RD.dRMST(S=S, W=W, horizon=5)
}
