% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_based_concordance.R
\name{mb.c}
\alias{mb.c}
\title{mb.c}
\usage{
mb.c(p.hat)
}
\arguments{
\item{p.hat}{Predicted probabilities}
}
\value{
The output of the mb.c function is:

mb.c

Model-based concordance
}
\description{
This function calculates the model-based concordance based on predicted probabilities. Please note this function works for predictions of binary outcomes only.
}
\examples{
library(PredictionTools)
set.seed(1)
n <- 100
p.hat <- runif(n, 0, 1)
PredictionTools::mb.c(p.hat=p.hat)
}
