% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_surv_mi.R
\name{val.surv.mi}
\alias{val.surv.mi}
\title{val.surv.mi}
\usage{
val.surv.mi(
  p,
  y,
  g = 5,
  time = NULL,
  main = "",
  lim = c(0, 1),
  dist = TRUE,
  CI = FALSE,
  df = 3,
  CI.metrics = FALSE,
  show.metrics = rep(TRUE, 4),
  optimism.C = 0
)
}
\arguments{
\item{p}{Matrix with predicted probabilities for imputation i in columns (complete case analysis: one column)}

\item{y}{Time to event outcome as Surv object (time,status)}

\item{g}{Number of risk groups; default=5}

\item{time}{Time point at which to evaluate the predicted probabilities, default=NULL (not entered), the maximum time point will be taken}

\item{main}{Plot label, default=""}

\item{lim}{limit, default=NULL}

\item{dist}{distribution, default=TRUE}

\item{CI}{plot confidence interval, default=FALSE}

\item{df}{degrees of freedom to compute confidence interval, default=3}

\item{CI.metrics}{plot confidence intervals of calibration intercept, calibration slope, and Harrell's C-index, default=FALSE. Note: Confidence interval for Uno's C-index is not displayed, since it is not implemented in survAUC.}

\item{show.metrics}{TRUE/FALSE vector of length 4 indicating if plot should show (1) sample size, (2) calibration intercept, (3) calibration slope, (4) Harrell's C-index possibly corrected with optimism specified in optimism.C, (5) Uno's C-index, default=rep(TRUE, 5)}

\item{optimism.C}{optimism-correction for Harrel's C-index in plot, default=0}
}
\value{
The output of the val_surv_mi function is a "list" with the following components.

main

Main title of plots.


n

number of observations


quants

quantiles


p.mi

predicted survival for each imputed data set.


obs.mi

observed survival for each imputed data set.


obs.mi.lower

lower bound of 95% confidence interval of observed survival.


obs.mi.upper

upper bound of 95% confidence interval of observed survival.


int

intercept


int.lower

lower bound of 95% confidence interval of intercept.


int.upper

upper bound of 95% confidence interval of intercept.


slope

slope estimate


slope.lower

lower bound of 95% confidence interval of slope.


slope.upper

upper bound of 95% confidence interval of slope.


cindex

C-index, uncorrected for optimism.


cindex.lower

lower bound of 95% confidence interval of C-index.

cindex.upper

upper bound of 95% confidence interval of C-index.


uno.C

Uno's C-index for time-to-event outcomes, uncorrected for optimism
}
\description{
This function calculates intercept, slope, and C-index for survival risk predictions of multiple imputed data set(s).
}
\examples{
library(PredictionTools)
library(survival)
library(rms)
library(mice)
set.seed(1)

# load data
data(pbc, package="survival")

# Multiple imputation
m <- 5
candidate.predictors <- c("sex", "age", "bili", "ascites", "hepato")
imp <- mice::mice(pbc[, candidate.predictors], m=m, print=FALSE)

# set survival time
S <- survival::Surv(time=pbc$time/365, event=as.numeric(pbc$status!=0))

# limit survival to time horizon
horizon <- 5
S.5 <- S
S.5[S[,1]>horizon, 1] <- horizon # if time is more than horizon, set to horizon
S.5[S[,1]>horizon, 2] <- 0       # if time is more than horizon, censor

p <- c()
for (i in 1:m){
 # fit model
  model.i <- cph(S.5 ~ sex + age + bili + ascites + hepato,
                 data=mice::complete(imp, i),
                 x=TRUE, y=TRUE, se.fit=TRUE)

  # obtain linear predictor
  lp.i <- predict(model.i, type="lp")

  # obtain baseline hazard
  f.basehaz.i <- survival::basehaz(model.i)

  # obtain baseline hazard at horizon
  h0.i <- f.basehaz.i$hazard[f.basehaz.i$time==max(f.basehaz.i$time[f.basehaz.i$time<=horizon])]

  # make predictions at time horizon
  p.i <- PredictionTools::fun.event(lp=lp.i, h0=h0.i)

  # add to p
  p <- cbind(p, p.i)
}
# internally validate predictions
g <- 4
main <- "Plot label"
show.metrics <- rep(TRUE, 5)
PredictionTools::val.surv.mi(p=as.matrix(p), y=S.5,
                             g=g, main=main, time=horizon,
                             show.metrics=show.metrics)
}
